Make FailedDocsHandler a typing.Protocol
We want users to be able to provide what they want here really, as long
as it matches the signature of the __call__ - so a Protocol makes more
sense, otherwise if users don't explicitly inherit from the
FailedDocsHandler class, mypy (and other type checkers) will complain.

Note that this allows a function to be used in place of a class if
desired.