docs: specify 'titles only' for API docs toctree
The default behaviour appears to have changed between Sphinx 5 and
Sphinx 6, which made the API docs include the toctree of each page (so
include all classes, functions etc.)

This was far too much to be useful - here we just want the titles of
the files to see the structure.

Adding this option means we can generate reasonable versions of the docs
with recent versions of sphinx and furo.