configure vulture to run on a weekly schedule
This would be too noisy for pre-commit: commit-by-commit you might
expect to add a function before it gets called anywhere, or have made
some code 'dead' but not removed it yet.

It's also fairly prone to false positives with kazu because of the
nature of how we use Hydra to actually instantiate our Pipeline(s),
and we don't want babysitting vulture to slow down 'in-the-moment'
development, it's better as a 'hmm, maybe there's some dead code I could
remove sometime' kind of task given the low urgency.