use dict literal rather than dict()
This seems easier to read/more obvious to me - I can just see 'what the
dict is' rather than having to metally convert the dict call to the
dictionary you would get in my head.

Note that Pylint has a code for 'use-dict-literal':
https://pylint.pycqa.org/en/latest/user_guide/messages/refactor/use-dict-literal.html

Although it partly justifies this with performance, which given this
won't get called much (and within the scope of running a ray server,
which is massively more expensive in comparison), isn't much of a
concern here, but is an added bonus.