remove pandas cast typing workaround
pandas-stubs used to be overly restrictive about the names argument,
meaning that our code worked fine (and was valid according the Pandas
docs in terms of the described accepted types), but pandas-stubs
believed it to be incorrect. We had a type cast to workaround this
previously.

pandas-stubs now accepts this type for the names argument (and a range
of other types), after
https://github.com/pandas-dev/pandas-stubs/pull/630 (by me).