move model training dependencies to optional group
seqeval in particular causes problems because it doesn't distribute a
wheel, only an sdist, which causes problems in environments with
proxies. pytorch_metric_learning doesn't cause this problem, but is also
a dependency specific to model training (and evaluation), so it makes
sense to make this optional too, so users who never intend to train
these model (which we expect to be nearly all users) don't have disk
space etc. taken up by these dependencies.

Style of handling ImportError matches what we already do for tensorboard
in pipeline.py