add config to blacken-docs
Blacken-docs doesn't pick up the black config in
pyproject.toml.

Leaving the line length at 88 rather than 100 as in the actual code
base, because a line length of 100 results in needing horizontal scroll
bars in (some) code snippets with our currents docs setup.

There also doesn't appear to be a config file option for
blacken-docs, so if/when we run blacken-docs locally, we'll also have to
remember these args which is unfortunate.