use new step_group functionality in server
Now we can use step_group for ner_only and linking_only (and also allow
specifying it in custom_steps), so that we get the groups from hydra in
a more principled way.

Although note that the names of the groups for the hard-coded ner_only
and linking_only endpoints are still 'hard-coded' as strings rather than
being looked up in config. We could use a 'string constant' or an Enum
or something, but it seems like as long as we test these endpoints, we
shouldn't be likely to remove/break these groups anyway.