add informational steps and step_groups apis
It's currently potentially hard for API-only users to find out which
step names and step group names are valid. Adding these APIs enables
them doing this without going looking in the kazu hydra config, which
would be hard for people outside of kazu development.

There's potentially some snazzy way of doing this so that the OpenAPI
spec gets all the valid values to show to the user. However, this might
be tricky (/potentially not doable), and porbably these API calls are
still desireable in that case, because people working on the command
line (possibly in a remote environment) may still want to be able to get
this information wihtout opening a webpage.

Another nicety would be to include in the API response for error
messages for requests specifying custom but invalid step or step_groups,
to include the valid options.