fixup Step docstrings
Many of the links to document_iterating_step and document_batch_step
were getting added to steps without a specific docstring, but the link
was broken for them, but we didn't see that because __call__ docstrings
weren't turned into Sphinx html.

Fix the links, plus ensuring 100 max line length where possible, and
link to Step.__call__ in a couple of places where we refer to it, but
couldn't previously link directly.