make ner_simple_test_cases a pytest fixture
We called this function in a few different places, and always did the
same thing with it - created Documents with the test cases' text.

Instead, use a fixture so we share this code (and the scope is clear,
that we only use within one function because Documents are mutable,
whereas that wouldn't be as clear if we just made the function also
create Documents).

Note that we don't currently use the entity class at all, but I'm
leaving this in in case it's useful data in the future, and its not
doing any harm.