refactor uri stripping
This used to be done while creating mappings in the MappingStep.
However, this previously wasn't configurable, and shouldn't really have
been coupled to the other logic of the mapping step (which is already
very complex).

Moving this logic to an Action on the CleanupStep, which makes it
configurable by users and easier to understand, and manages the
complexity better - we've simplified the MappingStep, and had to replace
that complexity with something that's very simple and has a single
concern.