rewrite get_request_id: for typing and decoupling
mypy compained that it didn't know what was inside the _list
attribute of the __dict__ of Headers. However, looking at this,
we're getting too close to the implementation of Headers anyway (e.g.
they could rename _list to _header and it would break us but while
maintaining their 'public' api).

Rewrite to use a 'public' method on Headers instead.