add typing for kazu_disk_cache
Diskcache doesn't distribute type hints, so the memoize decorator messes
up a bunch of type information here, even though it mostly just 'wraps'.

Unfortunately, there's no way I can find to do this preserving all
information and working correctly - in general, it looks like type
annotating decorators that can be used for methods or for functions
isn't really doable here, see https://github.com/python/mypy/issues/10805

Maybe this is doable in python 3.10+ or with a from __future__ import
so you can use ParamSpec and Concatenate or something. However, I this
point I was already frustrated trying a bunch of things and them not
quite working, and the current situation seems 'good enough'. We're
missing the types of the args to the decorated function (it now shows as
accepting anything), but we known the return type, and it won't produce
mypy errors that need to be ignored with every new use.