switch mypy setting to 'strict except for...'
Previously, we individually turned on a bunch of flags included in
'strict', and had manually taken out checks that were too noisy for us.

Switch to doing this declaratively, so we're saying 'do all the strict
checks, except for ...'. This doesn't change behaviour now, but means we
will pick up any new 'strict' checks in future,- and the intent is
clearer.