sort out sphinx comments at start of docstring
I had to change the docstring, because docformatter wants something
on the first line, but doesn't seem to recognise that this is Sphinx
markup, and the second line need different indentation, so just
shortening so it fits on one line. We could probably also write a
single-line summary to fix this problem.