switch on mypy's disallow-incomplete-defs
We had previously turned this off, but we've now fixed all the problems,
so can turn this back on.

However, turning off for tests specifically, since type hints here
aren't especially useful/necessary, and in particular, all pytest 'test'
functions return None, which would be tedious to have to annotate for
everything.