handle 'A.' and 'a.' tokenization
We want things like 'patients with Haemophilia A. Another sentence' to
split 'A.' into two tokens, 'A' and '.', so that our PhraseMatcher can
pick up 'Haemophilia A' in this sentence.

To do this, we need to remove the tokenizer exceptions that say things
like 'a.' that could be the start of a list should remain as one token
(the possibly of incorrect tokenization there isn't a concern for us).
We also need to ensure that '.' gets picked up as a 'suffix' when
preceded by a capital letter.

This requires a custom language class to get right (otherwise things
like 'a.' have already been added to the spacy vocab, which breaks the
behaviour). If we have that, it's also worth rolling the existing custom
tokenizer infix behaviour into that implementation, which is done as
part of this commit.