pin fastapi and pydantic - recent new releases break
Fastapi in particular - it breaks the way our 'examples' currently work
in the Swagger page.

This is because in the new version, FastAPI has used newer versions of
Json-Schema and OpenAPI Spec, which means the 'examples' can now be
'semantically correct', but then Swagger can't handle it properly, so
the experience is broken for end users of the Swagger web page, which is
what we really care about.

See https://fastapi.tiangolo.com/tutorial/schema-extra-example/?h=examples#technical-details
and https://github.com/tiangolo/fastapi/discussions/9784