remove Section.get_text and preprocessed_text
We only used this when we were using the abbreviation finder as a
preprocessing step, but we found that wasn't the best way of running
abbreviation finder, so preprocessed text has been None and get_text has
just been getting Section.text for a long while. Removing these
simplifies this without changing behaviour.