add upper limit to ray
HTTP Proxy Health Checks were added in ray 2.5, which by default kill
the job if the http proxy isn't up in 5 seconds. Kazu takes a lot longer
than this, and setting the environment variables to override this is a
bit of a pain for our setup - just downgrading for now until we decide
how exactly we want to solve this.