give TypeVars visible docstrings to prevent sphinx error
Sphinx tries to link to these when they are used in type annotations of
a function/class, so we get a warning (elevated to an error) if the
typevar doesn't have a reference in the docs.

To ensure there is a reference, we add a docstring, and make sure the
typevar isn't 'private' (rename _Ret to Ret).