turn on Sphinx 'python_use_unqualified_type_names'
Using builtin generics instead of those from typing currently causes
sphinx to behave differently such that the type parameters of the
generics, get long, 'fully qualified' versions of the type, which are
long, and make things much less clear - see this issue:
https://github.com/sphinx-doc/sphinx/issues/11571

Turning on this option 'cancels this out', and has almost no other
effect other than changing how things are italicised in some cases - I
don't think the new style of italicising is particularly worse (if
anything, it seems a little more consistent to me) so this seems like a
good outcome for now. We could consider turning this option off in
future if Sphinx resolves the issue above, but even in that case, it's
not necessarily causing any harm.