correct single backticks -> double backticks in rst
In rst, a single backtick causes italics, you need two backticks for a
code sample - see
https://www.sphinx-doc.org/en/master/usage/restructuredtext/basics.html#inline-markup
A single backtick results for us in a <cite></cite> pair of tags around
the text, which renders as italics for me. Fixing this - this is the
only occurence of this in the codebase that I could find.