fix tfidf hashing behaviour
python's builtin hash is non stable/consistent/predictable between
python processes, so the cache is 'missed' for us, so this is
essentially uncached at the moment.

Fix this by using hashlib, but this only takes bytes, so just pass
through the keys of the synonym data, which is all we actually use, so
that then, we can convert this set of string to a stable hash (sorting
is needed).