docs: no 'edit this page' button if it's autogenerated
Currently, there are 'edit this page' buttons even for the API docs,
which are autogenerated and aren't commited to the repo. Clicking this
links will give a 404 error, which is unpleasant for our users.

We add an autosummary/module template which adds `:generated:` to the
start of the autosummary generated .rst files. This is treated by
Rest/Sphinx as part of a 'meta' property for the page.

We also add a components/module template which extends the one from furo
so as to only include the furo button if 'meta' doesn't contain that
'generated' element (or doesn't exist at all). Therefore, the
autosummary api docs don't receive the 'edit this page' button.