docs: no 'edit-this-page' button for CHANGELOG
Because of the way we've had to set up the changelog doc build, this
would just take you to the docs/CHANGELOG.md page, rather than the
actual CHANGELOG.md page, so removing this link.

Perhaps we could mess about customizing where the furo edit button
points to for this page, so that it did edit the 'top-level'
CHANGELOG.md, but even then, we manage that file partly with towncrier,
and we wouldn't expect many edits here, so this seems like the best
thing to do for the effort.