ci: improve pytorch install process
Before, we tried to install everything from the pytorch 'index', and
then fail over to another index. This will likely be slow and wasteful,
as the pytorch index will fail for everything other than pytorch.
Instead, manually specify the pytorch install on an earlier line.

We didn't do this previously because it means the pytorch version in CI
isn't actually managed by pyproject.toml, so there isn't a single source
of truth. I've added a comment to pyproject.toml so this is at least
hopefully reasonably clear.