docs: make sphinx finish building and show all errors
Our -W option turns all sphinx warnings into errors, and by default
sphinx stops on the first error. This can prevent us from realizing that
there are actually multiple errors, and slowing down debugging
signficantly.

Instead, we can turn on '--keep-going' (which I didn't know existed
before!), so that the sphinx stil continues on until the end of the
build, so we see all the errors that exist.

Previously, I had been removing the '-W' in the Makefile locally when
testing, but often forgot this, and this won't be obvious to others less
familiar with sphinx.