docs: use default towncrier template
I discovered an issue with our existing template - it added trailing
whitespace whenever we had an  'orphan snippet' without a link to an
issue. Unfortunately for us, this is all our snippets.

We could have fiddled with the template, but I discovered at the same
time that in May of this year (2023), towncrier added a new default
markdown template, which hadn't existed when our 'custom' template was
added.

I believe our template was just intended to reproduce the
behaviour of the .rst template but in .md, so what the towncrier .md
template is doing - reading and comparing the two templates seems to
support this. The differences to me appear to be in that the 'build-in'
towncrier template supports features we don't use like 'sections', and
corrects the issue with trailing whitespace. In the absence of evidence
that we need a custom template ourselves, let's delete it.