ci: don't run dead code workflow on public repo
We already run it on our internal development repo, and want to use
self-hosted runners set up on that repo - currently the workflow fails
on our public repo because it doesn't have self-hosted runners.

We only need to run on one of the two repos, and the internal one seems
natural as development happens there first.

Approach follows:
https://docs.github.com/en/actions/using-workflows/workflow-syntax-for-github-actions#example-only-run-job-for-specific-repository