refactor: rewrite OpsinStep.extendString for efficiency
The style before meant calling it multiple times, and calling it with
spaces=2 first, which essentially does the computation for the spaces=1
result as part of its computation, so re-write to prevent
re-computation.

It also helps efficiency in cases where the match is at the end of a
section, whereby the old calls to self.extendString with spaces=2,
spaces=1 and spaces=0 would all have returned the same thing (if it
failed to parse), wasting computation. This also applies if it's one
'break' away from the end of the section, but with only spaces=2 and
spaces=1 providing the same result (spaces=0 would still be different).