docs: add changelog link for pypi
I noticed on pypi that some projects like
https://pypi.org/project/types-requests/ have a 'changes'/'history' etc.
link and a nice scroll icon. Adding for our pypi release.

This change won't take effect until the next pypi release.

I also managed to find someone's link to the code that decides on the
icon, to see if there were any other common links we might want, but it
didn't seem like it to me:
https://github.com/pypi/warehouse/blob/cd9dd933343f855aa8a3a8235c9baf46b0421249/warehouse/templates/packaging/detail.html#L20-L60