refactor: rework ent_fp_class_results type and similar
Key change is that this is now a single bool that with update
iteratively with 'or' so we don't have to keep all the values around. We
also renamed the type hint _KeyToMatcherResults which was
_EntityToClassResults, to better reflect what it's doing. Finally, we
changed these ent_fp_class_results and similar from defaultdicts to
plain dicts and used a .get instead.

Co-authored-by: Richard Jackson <richard.jackson4@astrazeneca.com>