mypy: make overrides for 'less strictness' module-specific
When we moved to using 'strict = true' for mypy, we turned off 4 of the
checks that mypy's strict mode includes because they impacted too much
of the existing codebase to change at that time. However, we did this
'globally', i.e. it affected all files that mypy checked.

Here, we switch to just turn off 2 of those checks on files where the check
is failing. This means the checks will be on for new files, which is
very useful. It also means the checks are on for existing files which
already passed the checks, which prevents us introducing issues there
too.

The 2 that we have left 'globally turned off' (disallow_untyped_calls
and no_implicit_reexport) are essentially ones where the source of the
problem is not where mypy says the error is (the error is where the
problematic thing is used), which means module-level overrides don't
really capture the things we want to ignore correctly. For both of
these, the benefit for the check seems quite small anyway.