Add type hints to make docsig and mypy happy
docsig doesn't like having a ':return:' in the signature description if
there isn't a return type annotation, so adding these. Note that I could
have just deleted the relevant ':return's, but this way, we actually
improve information for mypy too.

Once we'd added a return type hint though, our mypy setup then doesn't
like it if there's a return type annotation but not type annotation of
the other arguments, so adding these in where necessary.