docs: remove signature description from mock fixture
docsig didn't like that there's a ':return:' in the docstring but no
return type annotation. We could have added a return type annotation,
but it seems pointless here - it's a mock function for the tests. Also,
having a descriptor for 'monkeypatch' feels wrong when it's actually a
fixture request to pytest, not a real 'argument' in the normal sense.

None of this gets rendered anyway, so just delete the 'signature' bits
of this docstring.