docs: add missing param descriptions to __init__ methods
docsig doesn't check the __init__ docstring - it expects this to be
documented on the class if present.

I've opened https://github.com/jshwi/docsig/issues/195 , but I also
locally modified my install of docsig so that it did check these - this
is the first round of changes it spits up, where a param is just
missing from an __init__ signature description.