docs: fix problematic docstring with multiple lines
THis docstring has a literal lines break, which docformatter decides is
unecessary and removes, but it actually has semantic function in
separating the two parts of the tuple.

Sidestep this problem by rewriting the docstring to use a list for the
two elements of the tuple - docformatter is happy with this and its
clearer to my eyes anyway, both in the rendered docs but especially in
the actual raw docstring.