docs: upgrade docformatter and format
Upgrade the docformatter pre-commit hook to the most recent version.
This is necessary for us to be able to fix an issue with a Sphinx/ReST
'substitution' problem we currently have, without docformatter changing
it back to something broken (which is what broke it in the first place).

This is a large diff, but all changes are one of:

1. Re-wrapping lines (vast majority), which doesn't change the rendered docs
2. Capitalizing first letters of docstrings
3. Removing newlines after opening triples quotes of docstrings

All other changes that were introduced by upgrading docformatter I've
dealt with in the previous commits to highlight as different.