docs: remove erroneous double '\' from datamodel.rst
These was a double '\' because it was copied from data.py, but you only
need a single '\' in .rst - the double '\' is just what you have to do
in a python string literal to get an actual '\' in the string, because
python uses '\' as an escape character in strings, so you need two to
'tell python' "I'm not trying to escape the character after this, I'm
just trying to insert a literal backslash".

This meant that previously, a literal '\' was rendered in the docs,
which this change fixes.