refactor: calculate entity classes rather than needing them all
We need the entity classes so we can set the spacy token 'extensions' so
that the Matcher works as expected. We used to think we could get all of
these from the parsers, but this isn't true, because e.g. SETH can
produce entity matches that don't come from any parser, so then we
passed through an 'other_entity_classes' value for ones that don't come
from the parser. That seems a little messy and error-prone though - so
instead, we can just look at what entity classes are actually used in
the rules and that we therefore actually need to check.