update SpacyToKazuObjectMapper to fit previous commit
With the previous commit, we have a problem any time
__call__ encounters an entity with a class that isn't passed in to the
constructor within `entity_classes`. Add an arg to the __init__ so that
users can choose whether they want to just only set the extensions for
known entity classes, or use entity classes that occur in Sections
passed to __call__ as well as those passed in the constructor.

The later option is fairly error prone, as the user may try to read an
attribute that doesn't actually exist, because no section has (yet) been
passed that contains an Entity of the relevant type. As a result, we
also add signficant documentation of this risk, and also make the
riskier behaviour 'opt-in', and the safer behaviour the default.