rename _calculate_entity_classes_used_in_rules
Use _calculate_custom_extensions_used_in_rules since this is operating
over spaCy structures, and it's really all custom extensions that we're
checking here.

Rename related variables within the function as well, although the
class-level self.entity_classes_used_in_rules stays as-is,
referring to 'entity classes', since we want this to speak in the
language of the kazu data model, as users shouldn't have to know spaCy
to use the step.