ci: make version bump PR titles clearer
The PR titles were clear when they were ones like:

version bump: patch to v1.3.1

when the input version to the release workflow is a segment like
'patch', 'minor' etc.

But when an explicit version number like 'v1.3.0' is passed in, we get a
PR title like:

version bump: v1.3.0 to v1.3.0

Which I don't feel is obvious, and made me think the first number was
meant to be the old number and we'd made a mistake.

This changes it to look like:

version bump patch: v1.3.0 -> v1.3.1

or

version bump v1.3.1: v1.3.0 -> v1.3.1

which I think should be more obvious.