docs: fix broken reformatted docstrings
The recent version upgrade of docformatter reformatted this in a broken
manner (because there was no full stop at the end of the 'summary', and
no newline before the :param: 'field list').

Unfortunately I didn't notice this as part of
https://github.com/AZ-AI/kazu/pull/829 (although I noticed some other
similar cases).

Running docsig captured one of these - more positive signs that docsig is
useful!

The one in test_disambiguation_strategies.py I found using the regex
\S\s*:param to look for ':param' on a line where there's non-whitespace
characters before it on the line.