fix: don't force populate the parsers in explosion step
Now that the Explosion step is a ParserDependentStep, with the current
setup, we end up running the _populate_databases methods twice for it,
once when we call super().__init__() in the explosion step, but then
later, when building the phrasematchers, because we have 'force=True',
we repeat this.

I noticed this because if you build a model pack using the explosion
step and no curations, you can see the synonym generators running twice
for the same parsers in the output (which can take a considerable amount of
time!).

I can't see any reason to have force=True here anymore - the code is
older than our current caching strategy and I think ParserDependentStep,
so I think this is just something we missed an opportunity to update at
some point.