rename create_phrasematchers_using_curations
This name dated from a point where there were two methods to make the
phrasematchers - one from the parsers directly using the synonym
generators, and one using a file of curations.

The two functions are now unified - the 'split' between these two
options is now handled by the `populate_databases` method of
`OntologyParser` rather than in the OntologyMatcher.

As a result, the current name no longer really makes sense in referring
to the curations in the name.

Also fix up the docstring a bit while I was looking at it.