refactor: make GeneOntologyParser inherit from RDFGraphParser
This means much less duplication in the parsing code.

This came about because we wrote GeneOntologyParser before we had the
include/exclude entity patterns on the RDFGraphParser, so we used a
custom sparql query instead at the time. I didn't realise when adding
that feature on RDFGraphParser that it meant GeneOntologyParser could be
simplified.