chore: make docsig hook not 'local'
A local one means it's in the same venv as our actual install - but
docsig requires sphinx<7.0, which downgrades our sphinx for building the
docs and makes them look less good.

Having it 'non-local' means it's installed into its own venv, so then it
doesn't affect our sphinx version, and all is good!

This isn't one we should relaly need 'local' for, as there's not args
we're likely to want to pass running locally ourselves. I had it local
previously in part because I fiddled with the local install, but now it
has all the features that we need without fiddling with it locally.