ci: add missing torch install to dead code workflow
We never had this (oops!). We got away with this because we only run
vulture in this workflow, not the actual code.

However, it seems confusing that it seems like kazu is installed
properly here, but if you actually tried to run the pipeline, the torch
version would be one requiring a CUDA GPU, and would throw errors.