ci: move quick checks earlier in pr process
the pre-commit and docs build checks are relatively quick, but are often
the thing that fails in the integration tests. The reason we don't run
the full integration tests on all pushes to a PR/opening a PR is due to
the time taken to build the model pack, but we can check pre-commit and
the docs build without the model pack build.

The docs build causes particular pain, since it happens after the model
pack build in the integration tests, so would take a long time to fail
for potentially a simple to fix bit of broken ReST, which was very
annoying when it happened. We've also got caught out several times with
towncrier 'snippets' that don't end iwth a newline ending up on main,
which break pre-commit, which means they then need fixing as a priority
or all integration tests wil fail.

One caveat - currently this runs whenever a PR or opened, re-opened, or
updated. It's possible we ideally want to ignore PRs marked as draft,
but that involves more complex config, and to be honest I don't think
there's much wrong with running these checks on draft PRs anyway.