fix: pandas applymap function renamed to just map
The old name is deprecated, which meant that pandas-stubs removed the
type hint for it, which meant it got a fallback type hint that caused
mypy to complain.

However, I didn't want to just move to the new name, as this greatly
narrows our compatibility with pandas for the sake of a single line.

One downside is that it does mean an additional dependency of
'packaging' to parse the pandas package version, but we already get this
included in our transitive dependency graph by a bunch of different
routes (such as hydra, pytest, lightning among others), so this isn't
really adding a new dependency, just making it more visible.