docs: move label studio code to .py files
We have a longstanding issue to consider making these 'code-blocks' into
doctest blocks, so we can check the code is still valid. However, we've
never done this, because it's a faff with the labelstudio connection,
and getting the flow of the code right in the rendered docs but still
making sense for the tests.

However, if we just move the code out into a .py file, it then gets all
our good pre-commit hooks like black, flake8 and mypy. With mypy, we'd
generally know if this code isn't going to work for reasons like the
methods being refactored/renames/taking different types, so I think this
catches what we want.

One downside: the code used to be formatted by 'blacken-docs' rather
than black, which was actually good, because the black config of 100
line length introduces a horizontal scroll bar in the code blocks. To
resolve this, I've shortened a variable name so it fits without a scroll
bar!