tests: add server tests for paths
Now we have all the paths in routes.py , we can check we don't have
paths on the api that aren't added there too, and also, that the openapi
spec is reflecting the APIs that we expect to require auth correctly.

This will prevent us from making mistakes like like 'steps' and
'step_groups' where we added the apis, intended for them to not require
auth, but never added to routes.py to ensure that was the case.