webserver: add extra route for 'welcome to kazu' request
From the NO_AUTH_ENDPOINTS constant in routes.py , we can see this was
meant to cover this path, but didn't explicitly (I think the redirect
worked implicitly anyway). It's simple to add this, and not having it
meant that the test added in the previous commit failed, so adding this
now.