ci: use python 3.11 for github workflows
Also remove 'matrix' that we only ever used one value for.

Because the tests rely on a single built model pack, it would be tricky
to test with multiple versions of python, plus time consuming to run all
the pytest tests for multiple. The value of doing this seems fairly
small, and the costs reasonably high, so not worth expanding out the
matrix.

If we wanted to, we could always run a 'cronjob-style' workflow once
every couple of weeks/month to make sure other python versions can run
and pass the tests.