Add timeouts for label studio connection
Crucially a connect timeout, since the connection shouldn't take long to
establish, and could be the cause of timing out at the model pack build
level after 2 hours. We don't want to wait 2 hours to find out that the
label studio instance is completely dead and not responding, when if it
hasn't reponded for 3 seconds, it's almost certainly just down
completely.

Also add a read timeout to the 'project_id'
cached_property, as this API call to LabelStudio can be reasonably
'bounded' as this isn't hard to calculate and shouldn't be a big
response to pass through the network. By comparison, other methods like
deleting a project or adding a project depend on the size of the
project, so are less reasonable to attach an upper bound to.

On the plus side, adding a read timoeut for the 'project_id' actually
'protects' every other method apart from create_linking_project , as all
the others actually use self.project_id , so this also ensures we can
get some valid response from the server for that api call within a
reasonable period of time, which is good.