deps: add sphinx minimum version
For the previous commit to work, we need sphinx 7.2 or newer.

In theory, you could still build with Sphinx older than that, but you
would get a warning (which we turn into an error) for that type hint
specifically. Since it's a dev dependency, and there's no particular
reason for anyone to want to build with an older Sphinx version, pinning
seems like the more appropriate decision.