don't use cattrs class method strategy
This suffered from an issue in https://github.com/python-attrs/cattrs/issues/479
and ultimately meant 90% of the un/structuring code was in one place,
and then a tiny couple of functions elsewhere. Moving it all together is
simpler.

An added benefit is that this *might* be faster as well, since we can
use more cattrs generated un/structuring functions which should be
quicker.