fix: old Mapping.from_dict bug
comparing equivalence of this with the new cattrs method using
hypothesis brought up this bug. Fixing so the comparison complete
successfully.

WHen the "disambiguation_confidence" value was explicitly None,
previously, we weren't 'pop'ing it, which meant it stayed in
'mapping_dict', but then we were also explicitly passing through
disambiguation_confidence=disambiguation_confidence, so Mapping.__init__
received the same argument twice, which raises an error.

Fix this by always popping.