add SynonymTerm.from_dict
We had this for SynonymTermWithMetrics already, and it was just as easy.

In principle, we could have shared the definition here and just defined
it on the parent class SynonymTerm. But the type hinting would have been
weird, and given the function is only a single line it didn't seem worth
it.