Give SynonymTerm.mapping_types a default factory
Otherwise, when we serialize a document, the 'minify' logic removes
empty elements, which includes mapping_types, but the current cattrs code
doesn't handle adding back in some value for the mapping types.

We could customize the cattrs logic for this class, but to my eyes,
providing a default value within the cattrs code is a 'code smell' that
it should just be in the class all along.

Remove values of mapping_type where it's the default value anyway now.