feat: forbid 'extra' keys when structuring into data model
By default, cattrs is permissive and allows keys in the data you're
'structuring' that aren't fields in the target entities. However, this
risks typos for field names (which have default values) not being picked
up, and therefore the data loaded not being what was expected, so
'forbidding' it seems safer.

However, this highlighed that some 'structure' hook functions 'should'
have had `_cattrs_include_init_false=False` passed to them, fixing this
as part of the same commit