tests: introduce cattrs test converter to simplify tests
By including Entity._idx for the tests, we can directly compare Entity,
Section and Document round-tripping, and just have a short test to check
that the converter only differs in behaviour in the way we expect.

This also has the added benefit of more clearly (visually) encapsulating
the logic of setting up the converter in data.py.