tests: move hypothesis datamodel tests to test_data.py
These tests conceptually belong together, and over time, I expect to
have more and more hypothesis-based tests in the codebase, and I think
it makes sense for them to live alongside other pytest tests rather than
a separate folder.

It's possible that in future we'll want the strategies to live in the
conftest.py or somewhere so they can be shared with hypothesis tests in
other files, but for now, this seems fine.