fix: ensure Section hypothesis strategy can do sent spans
Previously, these were never generated, because _sentence_spans is
`init=False`, so they need to be 'added' after the init, but this isn't
really the responsibility of Hypothesis.builds /from_type normally. In
this case though, we do want it, as otherwise we haven't checked that
field can (de)serialize correctly.