refactor: move HFDataset to hf_token_classification.py
Previously this was in kazu.data.pytorch as the only item, which prevent
us from moving kazu.data.data to kazu.data.__init__ and just having a
single 'data' in our imports, which is nice. Given HFDataset only gets
used in this one file, it doesn't really justify the existence of the
kazu.data.pytorch file.