tests: generate optional args for registered hypothesis strategies
When we register a custom hypothesis strategy for a class, using
st.builds means that no optional arguments are passed at the 'top
level', only required arguments.

This previously meant we weren't really testing the datamodel fully.
Fortunately, there weren't any bugs in the bits we weren't testing
(which were mostly just standard out-of-the-box cattrs behaviour on top
of what we were already testing.

This does slow down the tests a little, because we're generating bigger
examples with deeper nesting of dataclasses.