test: convert pipeline test to use hypothesis
By using arbitrary generated text, we get much closer to 'test you can
run any text through the pipeline without it producing a processing
exception', which is a highly useful test for kazu.

The text currently uses the built-in hypothesis 'text' strategy, which
of course is not tailored to biomedical texts or NER/EL, so it is highly
plausible that there are very specific bugs that this text will fail to
find. This could be partly mitigated in future by a more specific
'strategy' for generating text, but this adds considerable value as-is.